вносит правки в отрисовку модального окна
